requirements.txt
alembic==1.4.2
backcall==0.2.0
bcrypt==3.1.7
btcpay-python==1.2.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
decorator==4.4.2
ecdsa==0.15
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.4.3
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jwt==1.0.0
Mako==1.1.3
MarkupSafe==1.1.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycoingecko==1.3.0
pycparser==2.20
Pygments==2.6.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
